Move functions to ReflectionExtensions
