+++ src/Scrutor/ReflectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
@@ -98,6 +99,75 @@ namespace Scrutor
             }
 
             return baseTypeInfo.IsAssignableToGenericTypeDefinition(genericTypeInfo);
+        }
+        
+        /// <summary>
+        /// Find matching interface by name C# interface name convention.  Optionally use a filter.
+        /// </summary>
+        /// <param name="t"></param>
+        /// <param name="action"></param>
+        /// <returns></returns>
+        internal static IEnumerable<Type> FindMatchingInterface(this TypeInfo t, Action<TypeInfo, IImplementationTypeFilter> action)
+        {
+            string matchingInterfaceName = "I" + t.Name;
+            var matchedInterfaces = GetImplementedInterfacesToMap(t).Where(x => string.Equals(x.Name, matchingInterfaceName, StringComparison.Ordinal));
+            Type type;
+            if (action != null)
+            {
+                var filter = new ImplementationTypeFilter(matchedInterfaces);
+                action(t, filter);
+                type = filter.Types.FirstOrDefault();
+            }
+            else
+            {
+                type = matchedInterfaces.FirstOrDefault();
+            }
+            if (type != null)
+            {
+                yield return type;
+            }
+        }
+
+        private static IEnumerable<Type> GetImplementedInterfacesToMap(TypeInfo typeInfo)
+        {
+            if (!typeInfo.IsGenericType)
+            {
+                return typeInfo.ImplementedInterfaces;
+            }
+            if (!typeInfo.IsGenericTypeDefinition)
+            {
+                return typeInfo.ImplementedInterfaces;
+            }
+            return FilterMatchingGenericInterfaces(typeInfo);
+        }
+
+        private static IEnumerable<Type> FilterMatchingGenericInterfaces(TypeInfo typeInfo)
+        {
+            var genericTypeParameters = typeInfo.GenericTypeParameters;
+            foreach (Type current in typeInfo.ImplementedInterfaces)
+            {
+                var currentTypeInfo = current.GetTypeInfo();
+                if (currentTypeInfo.IsGenericType && currentTypeInfo.ContainsGenericParameters && GenericParametersMatch(genericTypeParameters, currentTypeInfo.GenericTypeArguments))
+                {
+                    yield return currentTypeInfo.GetGenericTypeDefinition();
+                }
+            }
+        }
+
+        private static bool GenericParametersMatch(Type[] parameters, Type[] interfaceArguments)
+        {
+            if (parameters.Length != interfaceArguments.Length)
+            {
+                return false;
+            }
+            for (int i = 0; i < parameters.Length; i++)
+            {
+                if (parameters[i] != interfaceArguments[i])
+                {
+                    return false;
+                }
+            }
+            return true;
         }
     }
 }
